backend/app/utils/text_chunker.py
from langchain_text_splitters import RecursiveCharacterTextSplitter
from app.config.settings import get_settings


def get_text_splitter() -> RecursiveCharacterTextSplitter:
    settings = get_settings()
    return RecursiveCharacterTextSplitter(
        chunk_size=settings.chunk_size,
        chunk_overlap=settings.chunk_overlap,
    )