backend/requirements.txt
fastapi

uvicorn[standard]

langchain

langchain-community

langchain-nvidia-ai-endpoints

faiss-cpu

openai

tiktoken

pypdf

python-dotenv

pydantic

aiofiles

httpx

faiss-cpu

sentence-transformers

pymongo

langchain-nvidia-ai-endpoints